requirements.txt
google-genai
python-pptx
python-dotenv
lxml
flask
gunicorn